Split sync diagnose renderer
